intrinsic/evaluation/requirements.txt
allennlp==0.6.1
pytest
pytest-pythonpath
pytorch-transformers